Rewrote the app to use resume pdf instead of linkedin url
Completely Overhauled the app now it uses the resume uploaded as pdf, tweaked the webpage, used the gemini api properly and error-free, generating a webpage and downloading it as html works with no issues, occational images with incorrect source, but majorly production ready webpage deployed in seconds. Uses gemini-2.5-flash rather than gemini-2.0-flash, this takes longer but produces more visually appealing results. Worth the tradeoff. Obviuosly the preview gives a general idea of how the website will look, it is better visualised by downloading it and running the html in a browser.